fw-4435, move language factories
