+++ firstvoices/backend/tests/factories/site_info_factories.py
@@ -1,7 +1,7 @@
 import factory
 from factory.django import DjangoModelFactory
 
-from backend.models.sites import Language, LanguageFamily, SiteFeature, SiteMenu
+from backend.models.sites import SiteFeature, SiteMenu
 from backend.tests.factories.access import SiteFactory
 
 
@@ -12,21 +12,6 @@ class SiteFeatureFactory(DjangoModelFactory):
         model = SiteFeature
 
     key = factory.Sequence(lambda n: "Feature %03d" % n)
-
-
-class LanguageFamilyFactory(DjangoModelFactory):
-    class Meta:
-        model = LanguageFamily
-
-    title = factory.Sequence(lambda n: "Language Family %03d" % n)
-
-
-class LanguageFactory(DjangoModelFactory):
-    class Meta:
-        model = Language
-
-    title = factory.Sequence(lambda n: "Language %03d" % n)
-    language_family = factory.SubFactory(LanguageFamilyFactory)
 
 
 class SiteMenuFactory(DjangoModelFactory):
